@@ -24,9 +24,9 @@
     "@testing-library/jest-dom": "^5.11.10",
     "@testing-library/react": "^11.2.6",
     "@testing-library/user-event": "^12.8.3",
-    "amazon-cognito-identity-js": "^4.6.0",
-    "aws-amplify": "^3.3.27",
-    "aws-sdk": "^2.892.0",
+    "amazon-cognito-identity-js": "^4.6.3",
+    "aws-amplify": "^3.4.3",
+    "aws-sdk": "^2.1140.0",
     "axios": "^0.24.0",
     "d3": "^5.9.7",
     "d3-array": "^1.2.1",
@@ -45,11 +45,11 @@
     "d3-transition": "^1.1.3",
     "lodash.debounce": "^4.0.8",
     "papaparse": "^5.3.1",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
+    "react": "^18.1.0",
+    "react-dom": "^18.1.0",
     "react-query": "^3.28.0",
-    "react-router-dom": "^5.2.0",
-    "react-scripts": "4.0.3",
+    "react-router-dom": "^5.3.3",
+    "react-scripts": "^4.0.3",
     "react-table": "^7.7.0",
     "react-use-websocket": "^2.6.1",
     "recharts": "^2.1.9",
